requirements.txt
flask==3.0.0
flask-cors==4.0.0
anthropic==0.18.0
openai==1.12.0
python-dotenv==1.0.0
markdown==3.5.2
